[pattern-provider] 429 retry: fix first-event retry gap + surface WebStream HTTP errors properly
- Rewrote open_stream_with_retry to peek TWO events (Start + event_1)
  before committing. genai always emits ChatStreamEvent::Start first
  (from the SSE "Open" pseudo-event), so peeking only one event meant
  the retry loop never saw HTTP errors — they all arrived as event 1.
- Added WebStream downcast in map_genai_error: when WebStream.error
  downcasts to genai::Error::HttpError, map it to RateLimited (429) or
  RequestFailed (other status) instead of opaque RequestFailed{status:0}.
- Added WebStream downcast in server_rate_limit_hint for consistent
  Retry-After header extraction from tunneled HTTP errors.
- Added RetryThenSucceed custom wiremock responder (Arc<AtomicUsize>
  call counter) for stateful retry verification.
- Added anthropic_429_retries_then_succeeds test: verifies the retry
  loop fires and succeeds (call counter asserted to 2).
- Updated anthropic_429_surfaces_as_stream_error_without_content:
  removed .expect(1), updated assertions to handle exhausted-retry
  outcome correctly (RateLimited, not RequestFailed{status:0}).
- Added anthropic_oauth_first_party_beta_header_contains_both_markers
  regression test for the Critical 1 header-collision fix.

[pattern-provider] clippy: collapse nested if-let/match in WebStream error downcast

The new error-downcast logic in map_genai_error tripped collapsible_match
and collapsible_if lints. Collapse the two-arm pattern (downcast_ref Some
+ HttpError match) into a single if-let-Some destructuring the HttpError
fields directly.

Behaviour unchanged; passes existing 429 + 500 integration tests.

[pattern-provider] shaper reorg: separate Anthropic-specific from generic

Structure the shaper module tree so Anthropic-specific code lives under
a clearly-labelled namespace instead of masquerading as cross-provider
helpers at the shaper root:

  shaper.rs                   ← generic: RequestShaper trait, ShaperConfig,
                                ShapeContext, wrap_system_reminder
  shaper/
  ├── noop.rs                 ← NoOpShaper (moved out of shaper.rs)
  ├── anthropic.rs            ← HonestPatternShaper (moved out of shaper.rs)
  └── anthropic/
      ├── compat_mode.rs      (was shaper/compat_mode.rs)
      ├── headers.rs          (was shaper/headers.rs)
      └── system_prompt.rs    (was shaper/system_prompt.rs)

Public API is preserved via re-exports at shaper:: — downstream code
importing HonestPatternShaper, NoOpShaper, ShaperCompatMode, etc. from
pattern_provider::shaper works unchanged. The reorg is purely structural.

Motivation: the shaper trait + HashMap dispatch in PatternGatewayClient
are already per-provider, but free functions like build_beta_header_value
and types like ShaperCompatMode lived at the shaper module root with
provider-agnostic signatures. That made the Critical 1 header-collision
analysis harder than it should have been (the reviewer had to prove
that Anthropic-only markers didn't leak to non-Anthropic providers)
and would have invited future maintainers to casually gate new
behaviour on fields that happen to work for all current providers.

ShaperConfig still lives at the top level despite being Anthropic-biased
(compat_mode, target_is_first_party, all the capability toggles are
Anthropic-only). Renaming to AnthropicShaperConfig is a worthwhile
follow-up but touches ~20 call sites; doing it in a separate change
keeps this reorg purely mechanical. Noted as future cleanup in the
shaper module docstring.

Tests + clippy + doc + audit + pre-commit all clean.

+++ crates/pattern_api/src/responses.rs
@@ -126,12 +126,16 @@ pub struct ChatResponse {
     pub usage: Option<UsageInfo>,
 }
 
-/// Usage information for model calls
+/// Usage information for model calls.
+///
+/// Widths match `pattern_core::types::provider::{TokenCount, Usage}` — u64
+/// across the board. Anthropic's count_tokens surface is native u64, and
+/// cumulative counts from long-lived sessions can exceed u32::MAX.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct UsageInfo {
-    pub input_tokens: u32,
-    pub output_tokens: u32,
-    pub total_tokens: u32,
+    pub input_tokens: u64,
+    pub output_tokens: u64,
+    pub total_tokens: u64,
     pub model: String,
 }
